Refactor webhook org.
