+++ cmd/webhook-manager/app/util.go
@@ -29,8 +29,8 @@ import (
 	"k8s.io/klog"
 
 	"volcano.sh/volcano/cmd/webhook-manager/app/options"
-	"volcano.sh/volcano/pkg/admission/router"
 	"volcano.sh/volcano/pkg/client/clientset/versioned"
+	"volcano.sh/volcano/pkg/webhooks/router"
 )
 
 func registerWebhookConfig(kubeClient *kubernetes.Clientset, config *options.Config, service *router.AdmissionService, caBundle []byte) {
