+++ cmd/webhook-manager/main.go
@@ -30,9 +30,9 @@ import (
 	"volcano.sh/volcano/cmd/webhook-manager/app"
 	"volcano.sh/volcano/cmd/webhook-manager/app/options"
 
-	_ "volcano.sh/volcano/pkg/admission/jobs/mutate"
-	_ "volcano.sh/volcano/pkg/admission/jobs/validate"
-	_ "volcano.sh/volcano/pkg/admission/pods"
+	_ "volcano.sh/volcano/pkg/webhooks/admission/jobs/mutate"
+	_ "volcano.sh/volcano/pkg/webhooks/admission/jobs/validate"
+	_ "volcano.sh/volcano/pkg/webhooks/admission/pods"
 )
 
 var logFlushFreq = pflag.Duration("log-flush-frequency", 5*time.Second, "Maximum number of seconds between log flushes")
